Fixed some ToDos

+++ src/java/jpo/gui/TableSorter.java
@@ -180,7 +180,7 @@ public class TableSorter extends TableMap {
     public int compare(int row1, int row2) {
         compares++;
         for (int level = 0; level < sortingColumns.size(); level++) {
-            Integer column = (Integer) sortingColumns.elementAt(level);
+            Integer column = sortingColumns.elementAt( level );
             int result = compareRowsByColumn(row1, row2, column.intValue());
             if (result != 0) {
                 return ascending ? result : -result;
